Navbar
Menu is functional, if you want to test it you will need local bootstrap and jquery, or replace the links. For the time being I made it white, so make sure the site has some colored background before testing it. I've temporally changed the logo too, the header is stretched by it so make sure to download all of the files. The navbar is fully responsive so you can see how it works with the content.
Footer is not stretched right now, I'm still working on it's responsiveness.